Fix windows by using prebuilt Qt tarball instead of building from scratch
- As the 5.11.3 Qt tarball is missing from qts download site, you must find
and replace in the hunter directory the URL for the qt source. The easiest
is to set a new hunter root and binary_dir and generate easy_profiler. When
it errors out on 5.11.3 download change the URL in the Unpacked hunter files.
- Once you change the line EASY_PROFILER_NO_GUI line in CMakeList.txt to
build the gui on windows (instead of the Release-prefix hack) you let the generate
finish and then tarball up the hunter Install directory into the new
Qt prebuilt